Create README.md
As a Data Scientist at Ola, I'm working on a project to predict whether drivers will leave the company or not. Recruiting and keeping drivers has been a big challenge for Ola because many drivers leave frequently, and it's easy for them to switch to other ride-sharing services like Uber. To help tackle this problem, I'll be using data from 2019 and 2020 to analyze the characteristics of drivers and predict if they're likely to leave Ola in the future.